refactor(core): extract SkillManager and ToolDisplay modules
Extract two deep modules to separate concerns from Agent and main.tsx:

1. SkillManager (src/core/skill-manager.ts, +162 lines) — owns skill
discovery, initialization, loading, and tool restriction. Extracted from
Agent (4 fields + 8 methods). Agent now delegates to _skillManager.

2. ToolDisplay (src/cli/tool-display.tsx, +120 lines) — owns
ThinkingTagFilter, formatArgs, displayToolExecution, outputJson.
Extracted from main.tsx (~120 lines of display utilities). main.tsx
re-exports ThinkingTagFilter and formatArgs for backward compatibility.

Net: -207 lines (67 insertions, 274 deletions across 3 modified files)
agent.ts: 738 → ~640 lines
main.tsx: 541 → ~420 lines

1,288 tests pass, 0 fail. Typecheck clean. Lint clean (237 files).
